backend/secure_auth/requirements.txt
argcomplete==3.6.2
asgiref==3.11.1
Beaker==1.12.1
beautifulsoup4==4.12.3
blivet==3.10.0
blivet-gui==2.6.0
boto3==1.37.38
botocore==1.37.38
Brlapi==0.8.5
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distro==1.9.0
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dnf==4.23.0
docopt==0.6.2
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
gunicorn==25.3.0
humanize==3.13.1
idna==3.7
iso639==0.1.4
Jinja2==3.1.6
jmespath==1.0.1
langtable==0.0.68
libcomps==0.1.20
libdnf==0.74.0
louis==3.28.0
lxml==5.1.0
Mako==1.2.3
MarkupSafe==2.1.3
nftables==0.1
olefile==0.47
packaging==23.2
Paste==3.7.1
pexpect==4.9.0
pid==2.2.3
pillow==10.3.0
ply==3.11
productmd==1.45
protonvpn-cli==2.2.11
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pyenchant==3.2.2
PyGObject==3.48.2
PyJWT==2.12.1
pykickstart==3.52
pyOpenSSL==23.2.0
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.2.2
python-meh==0.51
python-pam==2.0.2
pythondialog==3.5.3
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2024.11.6
requests==2.31.0
requests-file==2.0.0
requests-ftp==0.3.1
rpm==4.19.1.1
s3transfer==0.11.5
selinux @ file:///builddir/build/BUILD/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.0.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.7.2
soupsieve==2.5
sqlparse==0.5.5
systemd-python==235
Tempita==0.5.2
urllib3==1.26.20
whitenoise==6.12.0
xkbregistry==0.3
